# src/app/api/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is Almond Backend API server documentation.For this sample, you can use the api key `special-key` to test the authorization     filters.",
		"version": "1.0.0",
		"title": "Almond Backend API",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "email@email.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "localhost:8080",
	"basePath": "/api",
...
	"schemes": ["http"],
	"paths": {
...
		"/roles": {
			"post": {
				"tags": ["roles"],
				"summary": "Create user roles for Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Roles"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			},
			"get": {
				"tags": ["roles"],
				"summary": "Get user roles for Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/roles/{id}": {
			"patch": {
				"tags": ["roles"],
				"summary": "edit specific user role by its id on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "role id",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful Operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			},
			"delete": {
				"tags": ["roles"],
				"summary": "delete specific user role by its id on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "role id",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful Operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
...
		"/me": {
			"get": {
				"tags": ["user"],
				"summary": "get user profile on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/people": {
			"get": {
				"tags": ["user"],
				"summary": "get user profiles on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/people/{id}": {
			"get": {
				"tags": ["schedules"],
				"summary": "Get a user by id on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			},
			"patch": {
				"tags": ["schedules"],
				"summary": "Edit a user on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
...
	},
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
	"definitions": {
...
		"Roles": {
			"type": "object",
			"required": ["title", "description", "resourceAccessLevels"],
			"properties": {
				"title": {
					"type": "string",
					"format": "Your Title Sample"
				},
				"description": {
					"type": "string",
					"example": "Description Sample"
				},
				"resourceAccessLevels": {
					"type": "string",
					"example": "Resource Access Level Sample"
				}
			}
		},
...
		"ApiResponse": {
			"type": "object",
			"properties": {
				"code": {
					"type": "integer",
					"format": "int32"
				},
				"type": {
					"type": "string"
				},
				"message": {
					"type": "string"
				}
			}
		}
	},
...
}